registeration and login

+++ INDIACom/Models/MemberDocumentModel.cs
@@ -2,8 +2,8 @@
 {
     public class MemberDocumentModel
     {
-        public int ID { get; set; }  // Primary Key (Auto-increment)
-        public int UserID { get; set; } // Foreign Key to User
+        public long ID { get; set; }  // Primary Key (Auto-increment)
+        public long UserID { get; set; } // Foreign Key to User
         public string Name { get; set; }
         public string DocumentType { get; set; } // e.g., "Resume"
         public string FilePath { get; set; } // Path where file is stored
